requirement.txt
absl-py==0.15.0
addict==2.4.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
contextlib2==21.6.0
cycler==0.10.0
einops==0.3.2
flatbuffers==1.12
fvcore==0.1.5.post20220119
gast==0.4.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna==3.3
imageio==2.9.0
iopath==0.1.9
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
ml-collections==0.1.0
mmcv==1.4.5
monai==0.7.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
olefile 
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
portalocker==2.3.2
protobuf==3.19.0
ptflops==0.6.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
seaborn==0.11.2
SimpleITK==2.0.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
thop==0.0.31.post2005241907
tifffile==2021.10.12
timm==0.4.12
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
yacs==0.1.8
yapf==0.32.0
